WIP: moved configuration logic into its own class
